Add files via upload
alright guys... kinda struggling to use github here. I worked on it some more and added an additional file so that when the user hits '4' in the menu it opens youtube and plays a video. BEWARE tho it starts creating an infinite amount of windows. If I run the code in the main java file it seems to run just fine. wont have time to debug until tomorrow but I wanted to keep you guys in the loop.